Moving class to custom package name for easier integration
