+++ src/main/java/eu/daxiongmao/geocode/kdtree/KDTree.java
@@ -24,7 +24,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
  */
 
-package geocode.kdtree;
+package eu.daxiongmao.geocode.kdtree;
 
 import java.util.ArrayList;
 import java.util.Collections;
