+++ src/test/java/eu/daxiongmao/geocode/ReverseGeoCodeSingletonTest.java
@@ -1,4 +1,4 @@
-package geocode;
+package eu.daxiongmao.geocode;
 
 import java.io.File;
 import java.io.FileNotFoundException;
@@ -11,6 +11,9 @@ import java.util.Set;
 import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
+
+import eu.daxiongmao.geocode.GeoName;
+import eu.daxiongmao.geocode.ReverseGeoCode;
 
 public class ReverseGeoCodeSingletonTest {
